+++ cli360monitoring/lib/servercharts.py
@@ -12,7 +12,7 @@ class ServerCharts(object):
     def __init__(self, config: Config):
         self.config = config
 
-    def create(self, serverId: str, metric: str = 'cpu', startTimestamp: float = 0, endTimestamp: float = 0):
+    def create(self, serverId: str, metric: str = 'cpu', startTimestamp: float = 0, endTimestamp: float = 0, dataAsJSON: bool = False):
         """Create a server metrics chart for the specified server id"""
 
         if not serverId:
@@ -24,7 +24,9 @@ class ServerCharts(object):
             return ''
 
         params = self.config.params()
-        params['output'] = 'png'
+
+        if not dataAsJSON:
+            params['output'] = 'png'
         # metrics can be one of [bitninja, cpu, httpd, mem, network, nginx, ping, process, swap, uptime, wp-toolkit]
         params['metric'] = metric
         if startTimestamp > 0:
@@ -34,7 +36,10 @@ class ServerCharts(object):
 
         response_json = apiGet('server/' + serverId + '/metrics', 200, self.config, params)
         if response_json:
-            if 'uri' in response_json:
+            if dataAsJSON:
+                print(json.dumps(response_json, indent=4))
+                return ''
+            elif 'uri' in response_json:
                 uri = response_json['uri']
                 print(uri)
                 return uri